Cloud Functions/https_measure_14day_asis/requirements.txt
functions-framework==3.*
pandas
google-cloud-bigquery
db-dtypes
flask